requirements.txt
streamlit
anthropic
openai
google-generativeai
python-dotenv
reportlab
python-docx
